Refactor code so HostPlatform.getExtensionPipePath is only called from a single place
